requirements.txt
Django==5.1.4
channels==4.2.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
daphne==4.2.1
redis==5.0.8
autobahn==23.6.2
djangorestframework
django-rest-knox
pillow
olefile
drf-yasg